llama-index/llm/requirements.txt
openai
python-dotenv
llama_index
llama-index-llms-openai
